refactor: mypage -> user 서비스/컨트롤러에서 관리
